requirements.txt
elasticsearch==7.12.0
grnhse-api
jmespath==0.10.0
oauth2client==4.1.3
prometheus_client==0.9.0
python-dateutil==2.8.1
six==1.14.0
urllib3==1.25.9

